Added reorder flag to dev_tests.sh

+++ tests/dev_tests.sh
@@ -156,28 +156,30 @@ echo ""
 
 echo "Time elapsed :" $[(end_time - start_time) / 60] "minutes"
 
-echo ""
+if [[ "$1" == "reorder" || "$2" == "reorder" ]]; then
+    echo ""
 
-echo "Reordering test list by time taken"
+    echo "Reordering test list by time taken"
 
-rm -f test_stats
-for t in ${tests[@]}; do
-    start_time=$(cat $t/start_time)
-    end_time=$(cat $t/end_time)
-    time_elapsed=$[end_time - start_time]
+    rm -f test_stats
+    for t in ${tests[@]}; do
+        start_time=$(cat $t/start_time)
+        end_time=$(cat $t/end_time)
+        time_elapsed=$[end_time - start_time]
 
-    echo "$t $time_elapsed" >> test_stats
-done
+        echo "$t $time_elapsed" >> test_stats
+    done
 
-sort test_stats -k2n > test_stats_sorted
+    sort test_stats -k2n > test_stats_sorted
 
-echo '#!/bin/bash' > test_list.sh
-echo "" >> test_list.sh
-echo "tests=(" >> test_list.sh
-while IFS= read -r line; do
-    test=$(echo $line | awk '{ print $1 }')
-    echo "    \"$test\"" >> test_list.sh
-done < test_stats_sorted
-echo ")" >> test_list.sh
+    echo '#!/bin/bash' > test_list.sh
+    echo "" >> test_list.sh
+    echo "tests=(" >> test_list.sh
+    while IFS= read -r line; do
+        test=$(echo $line | awk '{ print $1 }')
+        echo "    \"$test\"" >> test_list.sh
+    done < test_stats_sorted
+    echo ")" >> test_list.sh
+fi
 
 exit $exit_code
